requirements.txt
ldap3==2.2.4
python-ldap==3.2.0
rdflib==4.2.2
requests==2.18.1
